Remove messages_tx from TuiContext
Having mutable state in the TuiContext works for the TUI itself, but it makes it very hard to test because tests interfere with each other. By splitting it out, we can test in a much more functionality style. In 99% of cases we can just access messages_tx as a param of EventHandler::update, but for TemplatePreview I had to drill it down through a bunch of component constructors.