Fix missing preloaded request
Fix the latest request/response not showing up when the app is first started, or after a collection reload. This was only happening if the first recipe in the list was selected (index 0). The problem was we weren't calling on_select when the default value (0) was selected. When it was loaded from the DB it still wasn't pre-selected because it detected no change.

Fixed by adding a builder pattern so we can define on_select then call select_index in the constructor.

Closes #199