Implicitly support cursor events in selects
Previously, you had to manually call set_area for components that wrapped SelectState or FixedSelectState in order for cursor events to work with them. This is because they are only data containers and not responsible for rendering, but the view needs to know their screen area in order to pass them events. If you forgot to call set_area, they'd never receive cursor events.

This adds a Draw impl on SelectState and FixedSelectState, and removes state_mut() for both. This forces you to call draw on the select, which means the screen area will be tracked correctly. Since each usage of select can have a different display method (list, table, etc.), the draw prop is just a widget ready to be rendered. This allows the parent to do the presentation, but forces it to be routed through the child.