Refactor request building
Instead of building RequestRecord first then converting to reqwest::Request, it's now flipped around. This means we can offload all the the request building logic to reqwest, and just copy what it built when it's done. This is mildly helpful now for query param and authorization stuff, but it will become incredibly helpful for forms, where the construction is even more complex.