Add !json variant for structured bodies
- Add !json tag to `body` field of recipes
- !json bodies can contain any data. Strings will be treated as templates

An important design decision here is whether or not we should automatically set the `Content-Type` header. In the issue I wrote that we wouldn't, but during implementation I played around with both options. In the end I went with what was laid out in the issue. The pros and cons of setting it automatically:

Pros

- Convenient, reduces boilerplate

Cons

- It may be incorrect, e.g. if the user actually wants `application/geo+json`
- Implicit behavior can be confusing
- No way to un-set it (you could overwrite the header with an empty value but you can't omit it entirely)
- For more complicated types (e.g. forms), we may want to rely on the header to tells us more about the body, which is incompatible with the reverse

Another reason to leave it explicit for now is it's a non-breaking change to add the implicit behavior later.

Closes #242