Add !json variant for structured bodies
- Add !json tag to `body` field of recipes
- !json bodies can contain any data. Strings will be treated as templates
- `Content-Type` header will be set automatically based on body type

The original design in the ticket was to use have the user still provide `Content-Type`, but I realized there isn't much value in that. It just requires more work for the user, and introduces another thing we need to validate (body type vs `Content-Type`). If the user needs a subset of `application/json` they can still override it. Otherwise, providing it ourselves is just more convenient. The other motivating factor behind is forms. The machinery that reqwest gives us to set form data automatically sets the header, so I think it makes sense to follow that lead.

Closes #242