Use terminal editor for Edit Collection action
Instead of using the default system editor for `.yml` files when launching the Edit Collection action, we now defer to the `$VISUAL`/`$EDITOR` env vars. This means we will probably open a text-based editor like vim, where previously it was likely a GUI editor like VSCode. This could potentially be annoying but I think in most cases will be more convenient for users. If we get complaints we can make it configurable.

Closes #262