Allow escaping template keys
This is a major change to parsing, including replacing nom with winnow. While this is technically a breaking change because it changes the behavior of existing templates, in practice the number of templates that will be affected is likely zero. The escape sequences were very unlikely to be used previously.

This has some potential performance issues, because it involves cloning raw sections of text unnecessarily. This may lead to a lot of wasted work. There's room for optimizing the common path where no escape sequences are used. That will come in another PR though.

Closes #149