Eliminate EnumChain
This was an over-abstraction. Instead, we can just duplicate the EditCollection action on every action menu. Eventually we'll need to attach additional context to this action anyway, to open the collection file to the correct location. The abstraction included a TON of boilerplate that just was not worth the tradeoff of saving ~10 lines of code.