Add --insecure flag, only use config for TUI
Also remove the `config` target of the `slumber show` subcommand, because it's no longer relevant to the CLI. It's a bit odd to have the config file control the CLI. Generally CLIs are only configured through flags. Also this split will make it much easier to break the crate apart into multiple subcrates in the future.

This is a BREAKING change.