Use spacebar to toggle state
Table rows and folders used to be toggled via Submit (Enter). This adds an additional action Toggle (bound to Space by default) that applies to folder expand state and checkbox row state. This may be a bit of a confusing change because what previously toggled rows will now send a request instead, but I think it's worth it for the long term improvement.

This also means you can send a request from any tab in the recipe pane now, which has been a goal for a while.