Revert "Add --insecure flag, only use config for TUI"
This reverts commit 837f61f25ea3fe5e54abf32eb51868f372d6e527.

While this change may be desirable long-term, I'm reverting it for now to avoid the breaking change. The main goal of this was to enable crate-splitting, and I've given up on that goal for now so the breaking change doesn't bring much value. The original commit can easily be re-applied to get this functionality back.