Fix RequestStore::load_latest for non-success requests
Previously, when loading the most recent request for a recipe+profile, it would *only* look at requests in the DB. Now it loads the latest from the DB into the cache, then returns the most recent request in the cache. This means failed/in-progress requests, which are *not* in the DB, can be returned now.