Add requests_v2 and ui_state_v2 tables
This usurps the existing requests and ui_state tables. For requests_v2, iIt flattens all fields into DB columns. This makes it easier to browse and query data, and it also makes it easier to migrate in the future. Previously if we wanted to add a required field to RequestRecord or ResponseRecord, we would've had to keep an old version of the struct around to migrate. Now we can just add a DB column like a normal migration.

For ui_state_v2, it flattens the type name into its own column, and replaces MessagePack with JSON, since we have to pull in serde_json anyway. The JSON will also be easy to browse.

We have to keep the old schema, including rmp_serde and the Deserialize impls, around to support migrating. At some point in the future we can abandon those, but that should be after a reasonable amount of time once users have mostly migrated to the new version. I also left the old requests table around in case something goes wrong during a migration, so data isn't lost.