Store collection path as UTF-8
Previously the collection path was serialized as messagepack. This means it was stored as platform-specific bytes, with a header. This changes it to be stored just as UTF-8, which in practice should be fine for all cases. It eliminates rmp_serde as a dependency (except for migrations), and makes the logic much simpler.