Fix query toggle rows persisting out of sync
This upgrades persisted to 0.2.0, which changed Persisted to save on every mutation rather than just on drop. This prevents bugs like the query toggle rows not saving when switching profiles. The issue occurs when a new Persisted value is created before the old one is dropped. By saving immediately when the mutation is made, it prevents this. Note that this change was *not* made to PersistedLazy because it's much harder to scope down the &mut lifetime to just when the persisted data is modified. I don't think the bug actually occurs anywhere in Slumber with PersistedLazy, but it's still possible.