Hide Collection behind Arc
Previously we cloned the entire collection for every template render. While I haven't noticed a particular performance issue with this, I also haven't worked with very large collections. This should be a pretty low-risk optimization, because the Arc is only ever cloned when passed between tasks, when the big clone would've happened previously.