Skip draws when UI falls behind
If we're getting a lot of inbound events (e.g. while scrolling) and the UI is struggling to keep up, we'll start skipping draws until the message/input queue is empty. This means the UI will start to look a bit jittery. The current state is that we'll queue up a bunch of outdated draws, meaning when the user stops doing their repeated input, they'll be locked out of interaction until the queue is drained. jittery > unusable