Cache highlighted text bodies
Previously, we were re-calculating syntax highlighting, and generating Text objects, for large bodies on every render. This applied to the recipe body, request body, and response body panes. For large bodies, this led to significant overhead. For a ~182KB body, the response pane went from ~100ms render time down to ~6ms.

This comes at the cost of some additional up-front clones, but I think the tradeoff is worth it.