Remove selected profile ID plumbing
This removes the selected_profile_id field from Message::TemplatePreview, which removes the need to plumb it all over the view. Instead, the main TUI loop will grab the currently selected profile from the view when it starts rendering the preview. This *does* introduce the potential for sync bugs, where the selected profile changes between when the preview message is queued and dequeued. In reality that should(?) be impossible, because the TUI drains all messages in the queue between draws. It's theoretically possible is most likely impossible in practice. The consequence is we just show some stale data in the UI so it's not the worst outcome.