[BREAKING] Don't deduplicate backslashes during parsing
The old backslash syntax meant raw:template syntax wasn't 1:1. It was possible to write two different strings that parsed to the same template. This means mapping back to strings wasn't necessarily an identical round trip. The new syntax is a breaking change, but will probably not affect most users, and should be an improvement for those who are impacted.