Make raw recipe bodies editable
Since bodies can be large, we provide editability through the external editor instead of the inline text box. Currently the entire body is read back from the file and stored in memory. It'd be possible to store the override as a text diff from the original, but probably not worth it. We'd still have to (1) read the entire new body from the file and (2) parse it as a template. Once those two things are done, keeping the new template around in memory it's that expensive.