Remove message batching optimization
The optimization was to skip draws while there are messages in the queue. The goal of it was to prevent the TUI from falling behind when draws get slow. The problem is this changes the semantics of the program. A draw updates the set of which components are visible, which affects event handling since only visible components are given events. This specifically caused a bug in the following chain of events:

1. User input triggers event (open a modal)
2. Event handling calls a component constructor (history modal)
3. Constructor queues a message (load history from DB)
4. Component is rendered immediately after construction, so it's marked as visible = true
5. Message contains a callback to queue a local event with the loaded data

With the optimization, step 4 is skipped because of the message queued in step 3. This means by the time the local event is handled in step 5, the new component was never rendered so it's still marked as visible=false and doesn't get to see its own event.

In general optimizations should not change the semantics of a program. In the future I think the best way to prevent the event loop from falling behind is to fix slow renders rather than skipping them. This may lead to some worse behavior for the user, but the tradeoff is less complexity in the main loop. This was a hard bug to find and I don't want to do it again.