Fix release by tweaking changelog
That's an interesting commit message, isn't it. Apparently having the words "Basic" and "Bearer" in the changelog causes the CI to identify the release plan as containing "secrets", and therefore doesn't share it from the plan step to the execution steps. This error messages around this from Github are incredibly poor so I had to find the offending words with some binary search-type thing.