Move RequestStore to top level of TUI state
This moves RequestStore out of the view and into root state. This means we can use it to track in flight requests. In the future we'll use it for cancellation tokens too.

I'm not particularly fond of this code, especially the Rc<RefCell<_>> in TestHarness. I'll probably clean it up eventually but right now I just need to make progress.