Persist response queries separately for each content type
This prevents running the command on the wrong response type when the response content type changes. E.g. if a response type changes from JSON to an HTML error page, we won't try to run the previous `jq` command on the body. When switching back to JSON, the old command will come back.