Use enum for representing query param values
Instead of support a list of param=value strings, we now support a map of {param: value} OR {param: [value, ...]}. This avoids string parsing, is easier to deal with internally,and is more consistent with the rest of the collection format.

This is obviously a breaking change. It will be bundled with other breaking changes and an importer will convert from the old format to the new one.