Fix fx opening slowly sometimes
This was a complicated interaction between slumber and fx related to event handling on the tty. The subprocess is spawned in the main thread and blocks the entire thread, which prevents slumber from doing any work (this is intentional). The issue shows up when the event reader stream is reading for an event (which happens in a background task) while the subprocess is spawned. It ends up contending with fx during startup. The issue happened intermittently because by default, tokio::select! randomly chooses the order in which branches will be polled. If the event reader branch was polled first (one third of the time), the issue happened.

This is a simple and small fix: always check the message queue _before_ spawning the event reader. If a message is immediately available, the event reader won't spawn. Since the subprocess spawn is triggered by a FileView message, this prevents the event reader from spawning before the subprocess.