Fix flaky http tests and improve test perf
A reqwest client can't be used across multiple tests because each client is bound to a particular tokio runtime and #[tokio::test] creates a new runtime for each test (see https://github.com/tokio-rs/tokio/issues/2374).

The fix for this is to create a new HttpEngine for each test, which will create a new reqwest client as well underneath. This fixes the flakiness but adds a 100-300ms delay to each test.

I fixed the delay by disabling loading native certs in tests only. We need to keep the native cert loading enabled in prd because of https://github.com/LucasPickering/slumber/issues/275