Use terminput for input defs in config crate
Eliminate's config crate's dependency on crossterm. terminput is smaller so the config crate has a smaller tree now. This probably increases compile time overall though because it's one more crate to compile. This will enable compilation for the web in the future though because terminput supports the web-sys backend as well.

This required updating to ratatui@0.30.0-alpha.5 in order to get everyone on crossterm 0.29.