Skip terminal update if when there are more updates
If we're doing a draw and there's more messages/inputs in the queue, we can skip the updates to the terminal. We still need to call all the internal draw logic because it also updates state, but when we're making multiple draws together, we don't need to update the terminal until the last one. This improves performance when sending a lot of rapid input, e.g. holding down an arrow key.