Skip disabled items in SelectState
It's now impossible to select a disabled item in a select state. This is useful for tabs since they now support disabling items, but it's also more useful for select lists because there's nothing you can do with the selected item anyway. I didn't do this originally just because it's more complicated and potentially buggy. I wrote a prop test though so it can't possibly be broken!!