Add jq template function
jq is much more powerful than jsonpath. This uses jaq which is a pretty faithful reimplementation of jq in pure Rust. I investigated using jq-rs instead which links to libjq, but it required additional tools to install which I don't want because some people want to do `cargo install slumber`.