Accept labelled object in select()
select() now accepts an array of objects like {label: "Label", value: 1} (in addition to the existing support for an array of strings). This allows you to specify one value for display but return a different value (e.g. select a user by name, then forward on the ID). It also allows select to output arbitrary values instead of just strings.

Right now the easiest way to use this is to pipe to command() and call jq. I plan to add a jq() function to support jq natively, which will make it even easier.

Closes #609