Convert Component into a trait
This is a major refactor of the TUI framework. Component is now a trait instead of a wrapper struct. EventHandler has been merged into Component. Each component is now resposible for storing its own ID but the render machinery is otherwise managed automatically. Having a component be aware of its own ID will make it possible simplify event handling in the future.