Use static YAML file for default config
Instead of dynamically generating the default config file from Config::default(), we now use a static default.yml file. The main reason is to avoid populating the file with ALL fields. While this may be useful to users, it makes it much harder to change default values in the future. Those changes would only be rolled out to new users, and existing users with the default file would continue to have the old values. This is important because I plan to change some default keybindings in version 5.

There's also a step that detects the old default file and replaces it with the new one. This will undo as much of this mistake as possible. It won't replace any file that's been modified by the user in any way, so there's minimal risk of it overwriting something the user cares about.