Route TemplatePreview render through event queue
Instead of using Mutex to write the result of template preview, this routes it through the event queue instead. The mutex was a pretty old solution and became inconsistent with the rest of the app architecture. Using an event also means we don't have to send Message::Draw from every spawned task. It was the only task that actually needed it.

For now I'm emitting the callback event directly from the background task. This works because the task is run on the main thread, as it's in a LocalSet. It's possible we may want to push some or all of the work into a background thread because it may be slow for large bodies. In this case we'll need to route the callback through a Message instead.