Use strings for override keys
Simplify the maps used to override headers/query params/form fields to use strings. In the case of query params, strings along aren't unique, so we also use key-specific indexes. This will make it much easier to define these overrides via the CLI. The index-only keys was only good for the TUI.